% Auto-generated: do not edit by hand
\name{aspectRatio}

\alias{aspectRatio}

\title{AspectRatio component}

\description{
Aspect Ratio component
}

\usage{
aspectRatio(children=NULL, id=NULL, ratio=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{ratio}{Numeric. Element ratio}

\item{styleProps}{Named list. StyleProp object
Use maxWidth to adjust conent width}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

