% Auto-generated: do not edit by hand
\name{box}

\alias{box}

\title{Box component}

\description{
Box Component
}

\usage{
box(children=NULL, id=NULL, asProp=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{asProp}{Character. Render element as tag}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

