% Auto-generated: do not edit by hand
\name{breadcrumb}

\alias{breadcrumb}

\title{Breadcrumb component}

\description{
Breadcrumbs, or a breadcrumb navigation, can help enhance how users navigate to previous page levels of a website, especially if that website has many pages or products. Composes to `Box`, so all its props can be passed via `styleProps`
}

\usage{
breadcrumb(children=NULL, id=NULL, separator=NULL, colorScheme=NULL,
size=NULL, spacing=NULL, variant=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{separator}{Character. Breadcrumbs separator
Currently supports only strings}

\item{colorScheme}{Character. Color scheme}

\item{size}{Character. Breadcrumb size
Not implemented in default theme}

\item{spacing}{Character | numeric. The left and right margin applied to the separator}

\item{variant}{Character. Breadcrumb variant
Not implemented in default theme}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

