% Auto-generated: do not edit by hand
\name{breadcrumbItem}

\alias{breadcrumbItem}

\title{BreadcrumbItem component}

\description{
BreadcrumbItem element Does not provide Link
}

\usage{
breadcrumbItem(children=NULL, id=NULL, isCurrentPage=NULL,
isLastChild=NULL, spacing=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{isCurrentPage}{Logical. When `true`, the BreadcrumbLink renders a span
with `aria-current` set to page instead of an anchor element.
Defaults to false}

\item{isLastChild}{Logical. Is last child?
Defaults to false}

\item{spacing}{Character | numeric. The left and right margin applied to the separator}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

