% Auto-generated: do not edit by hand
\name{center}

\alias{center}

\title{Center component}

\description{
Center component
}

\usage{
center(children=NULL, id=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

