% Auto-generated: do not edit by hand
\name{chakraProvider}

\alias{chakraProvider}

\title{ChakraProvider component}

\description{
Chakra UI context provider
}

\usage{
chakraProvider(children=NULL, themeExtension=NULL, resetCSS=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{themeExtension}{Named list. Object to extend Chakra UI theme}

\item{resetCSS}{Logical. Reset CSS inside provider}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

