% Auto-generated: do not edit by hand
\name{code}

\alias{code}

\title{Code component}

\description{
Code component
}

\usage{
code(children=NULL, id=NULL, colorScheme=NULL, size=NULL,
variant=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{colorScheme}{Character. ColorScheme
Defaults to 'gray'}

\item{size}{Character. Code component size}

\item{variant}{Character. Code variant
Defaults to 'subtle'}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

