% Auto-generated: do not edit by hand
\name{container}

\alias{container}

\title{Container component}

\description{
Container Component
}

\usage{
container(children=NULL, id=NULL, centerContent=NULL,
colorScheme=NULL, size=NULL, variant=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{centerContent}{Logical. Center content}

\item{colorScheme}{Character. ColorScheme
Not implemented in the default theme}

\item{size}{Character. Component size 
Not implemented in the default theme}

\item{variant}{Character. Component variant
Not implemented in the default theme}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

