% Auto-generated: do not edit by hand
\name{flex}

\alias{flex}

\title{Flex component}

\description{

}

\usage{
flex(children=NULL, id=NULL, flexProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{flexProps}{Named list. FlexBox properties
For a list, refer to https://chakra-ui.com/docs/layout/flex#props}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

