% Auto-generated: do not edit by hand
\name{formControl}

\alias{formControl}

\title{FormControl component}

\description{

}

\usage{
formControl(children=NULL, id=NULL, label=NULL, helperText=NULL,
isRequired=NULL, size=NULL, variant=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{label}{Character. Label}

\item{helperText}{Character. Helper text}

\item{isRequired}{Logical. Is this form component required?}

\item{size}{Character. Component size}

\item{variant}{Character. Component variant}

\item{styleProps}{Named list. StyleProps}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

