% Auto-generated: do not edit by hand
\name{grid copy}

\alias{grid copy}

\title{Grid copy component}

\description{

}

\usage{
grid copy(children=NULL, id=NULL, gap=NULL, templateColumns=NULL,
templateRows=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{gap}{Numeric. gridGap}

\item{templateColumns}{Character. gridTemplateColumns}

\item{templateRows}{Character. gridTemplateRows}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

