% Auto-generated: do not edit by hand
\name{grid}

\alias{grid}

\title{Grid component}

\description{
Grid component Accept other components or `GridItem` as children.  For shortness, Grid props are passed in an object/dictionary. `Grid` composes `Box`, so any styleProps can be passed via `gridProps` For details, see https://chakra-ui.com/docs/layout/flex
}

\usage{
grid(children=NULL, id=NULL, gridProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{gridProps}{Named list. Grid properties (includes styleProps)
For a list, refer to https://chakra-ui.com/docs/layout/grid#props-1}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

