% Auto-generated: do not edit by hand
\name{gridItem}

\alias{gridItem}

\title{GridItem component}

\description{
GridItem component To be used as child of Grid component See: https://chakra-ui.com/docs/layout/grid#spanning-columns
}

\usage{
gridItem(children=NULL, id=NULL, colSpan=NULL, colStart=NULL,
colEnd=NULL, rowSpan=NULL, rowStart=NULL, rowEnd=NULL,
styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{colSpan}{Numeric. Number of columns that the item should span}

\item{colStart}{Numeric. Column number that grid item should start}

\item{colEnd}{Numeric. Column number that grid item should end}

\item{rowSpan}{Numeric. Number of rows that the item should span}

\item{rowStart}{Numeric. Row number that grid item should start}

\item{rowEnd}{Numeric. Row number that grid item should end}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

