% Auto-generated: do not edit by hand
\name{icon}

\alias{icon}

\title{Icon component}

\description{
Icon component that render icons from Chakra UI library If 'icon' does not exist in library, Chakra renders the fallback Icon (see https://chakra-ui.com/docs/media-and-icons/icon#fallback-icon)  For a list of icons, see https://chakra-ui.com/docs/media-and-icons/icon#all-icons
}

\usage{
icon(id=NULL, icon=NULL, viewBox=NULL, boxSize=NULL, color=NULL,
isFocusable=NULL)
}

\arguments{
\item{id}{Character. Component Id}

\item{icon}{Character. Chakra UI icon name
For a list, refer to the documentation
https://chakra-ui.com/docs/media-and-icons/icon#all-icons}

\item{viewBox}{Character. Icon viewBox
Defaults to '0 0 24 24'}

\item{boxSize}{Character. Icon boxSize
Defaults to '1em'}

\item{color}{Character. Icon color}

\item{isFocusable}{Logical. Is an interactive element or just for presentation}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

