% Auto-generated: do not edit by hand
\name{input}

\alias{input}

\title{Input component}

\description{
Text Input component
}

\usage{
input(id=NULL, value=NULL, placeholder=NULL, type=NULL,
isInvalid=NULL, colorScheme=NULL, size=NULL, variant=NULL)
}

\arguments{
\item{id}{Character. Component Id}

\item{value}{Character. Internal value parameter (DO NOT USE as input)}

\item{placeholder}{Character. Text placeholder}

\item{type}{Character. Input type
Defaults to text}

\item{isInvalid}{Logical. Is invalid}

\item{colorScheme}{Character. Color scheme}

\item{size}{Character. Button size
Defaults to 'md'}

\item{variant}{Character. Input variant 
Options are 'outline', 'filled', 'flushed' or 'unstyled'
Defaults to 'outline'}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

