% Auto-generated: do not edit by hand
\name{select}

\alias{select}

\title{Select component}

\description{
Select component
}

\usage{
select(id=NULL, options=NULL, value=NULL, placeholder=NULL,
isInvalid=NULL, isDisabled=NULL, isRequired=NULL, size=NULL,
variant=NULL, colorScheme=NULL)
}

\arguments{
\item{id}{Character. Component Id}

\item{options}{Unnamed list. Options}

\item{value}{Character. Selected value}

\item{placeholder}{Character. Text placeholder when
no option is selected}

\item{isInvalid}{Logical. Is the component invalid?}

\item{isDisabled}{Logical. Is the component disabled?}

\item{isRequired}{Logical. Is the component required?}

\item{size}{Character. Component size}

\item{variant}{Character. Component variant}

\item{colorScheme}{Character. Color Scheme}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

