% Auto-generated: do not edit by hand
\name{slider}

\alias{slider}

\title{Slider component}

\description{
Slider component
}

\usage{
slider(id=NULL, value=NULL, min=NULL, max=NULL, step=NULL,
isInvalid=NULL, isDisabled=NULL, isRequired=NULL,
orientation=NULL, size=NULL, variant=NULL, colorScheme=NULL,
styleProps=NULL)
}

\arguments{
\item{id}{Character. Component Id}

\item{value}{Numeric. Selected value}

\item{min}{Numeric. Minimum value}

\item{max}{Numeric. Maximum value}

\item{step}{Numeric. Step value}

\item{isInvalid}{Logical. Is the component invalid?}

\item{isDisabled}{Logical. Is the component disabled?}

\item{isRequired}{Logical. Is the component required?}

\item{orientation}{Character. Component orientation}

\item{size}{Character. Component size}

\item{variant}{Character. Component variant}

\item{colorScheme}{Character. Color Scheme}

\item{styleProps}{Character. StyleProps}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

