% Auto-generated: do not edit by hand
\name{spacer}

\alias{spacer}

\title{Spacer component}

\description{
Spacer component To be used as Flex child  See https://chakra-ui.com/docs/layout/flex#using-the-spacer
}

\usage{
spacer(id=NULL)
}

\arguments{
\item{id}{Character. Component Id (dummy)}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

