% Auto-generated: do not edit by hand
\name{square}

\alias{square}

\title{Square component}

\description{
Wrapper for Center component in square format
}

\usage{
square(children=NULL, id=NULL, size=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{size}{Character | numeric. Component size}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

