% Auto-generated: do not edit by hand
\name{stat}

\alias{stat}

\title{Stat component}

\description{
WIP This component is EXPERIMENTAL and might become deprecated. Show statistics in a card Composes to `Box`, so all its props can be passed via `styleProps`
}

\usage{
stat(id=NULL, statLabel=NULL, statNumber=NULL, statHelpText=NULL,
size=NULL, colorScheme=NULL, variant=NULL, styleProps=NULL)
}

\arguments{
\item{id}{Character. Component Id}

\item{statLabel}{Character. statLabel}

\item{statNumber}{Character | numeric | dash component. statNumber}

\item{statHelpText}{Character. statLabel}

\item{size}{Character. Stat component size 
Defaults to 'md'}

\item{colorScheme}{Character. ColorScheme}

\item{variant}{Character. Stat variant}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

