% Auto-generated: do not edit by hand
\name{switch}

\alias{switch}

\title{Switch component}

\description{

}

\usage{
switch(children=NULL, id=NULL, isChecked=NULL, isDisabled=NULL,
isRequired=NULL, size=NULL, value=NULL, variant=NULL,
styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{isChecked}{Logical. Is checked?}

\item{isDisabled}{Logical. Is disabled?
Defaults to false}

\item{isRequired}{Logical. Is checkbox required?
Defaults to false}

\item{size}{Character. Switch size
Defaults to 'md'}

\item{value}{Numeric | character. Value to be returned to forms}

\item{variant}{Character. Switch variant
Defaults to 'solid'}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

