% Auto-generated: do not edit by hand
\name{textarea}

\alias{textarea}

\title{Textarea component}

\description{
Text Textarea component
}

\usage{
textarea(id=NULL, value=NULL, placeholder=NULL, isDisabled=NULL,
isInvalid=NULL, colorScheme=NULL, size=NULL, variant=NULL)
}

\arguments{
\item{id}{Character. Component Id}

\item{value}{Character. Internal value parameter (DO NOT USE as input)}

\item{placeholder}{Character. Text placeholder}

\item{isDisabled}{Logical. Is disabled}

\item{isInvalid}{Logical. Is invalid}

\item{colorScheme}{Character. Color scheme}

\item{size}{Character. Textarea size
Defaults to 'md'}

\item{variant}{Character. Textarea variant
Options are 'outline', 'filled', 'flushed' or 'unstyled'
Defaults to 'outline'}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

