% Auto-generated: do not edit by hand
\name{wrapItem}

\alias{wrapItem}

\title{WrapItem component}

\description{
WrapItem component To be used as child of `Wrap` component (if `shouldWrapChildren` is `false`) Composes to `Box`, so all its props can be passed via `styleProps`
}

\usage{
wrapItem(children=NULL, id=NULL, styleProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{styleProps}{Named list. StyleProp object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

