% Auto-generated: do not edit by hand
\name{flex}

\alias{flex}

\title{Flex component}

\description{
FlexBox component Accept other components as children.  For shortness, Flex props are passed in an object/dictionary. `Flex` composes `Box`, so any styleProps can be passed via `flexProps` For details, see https://chakra-ui.com/docs/layout/flex
}

\usage{
flex(children=NULL, id=NULL, flexProps=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. Component Id}

\item{flexProps}{Named list. FlexBox properties (includes styleProps)
For a list, refer to https://chakra-ui.com/docs/layout/flex#props}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

